Decouple batch and legacy methods in stats collection
